user-controller

+++ src/main/java/com/swapnixi/feedbackhub/service/UserService.java
@@ -1,14 +1,15 @@
 package com.swapnixi.feedbackhub.service;
 
+import com.swapnixi.feedbackhub.entity.UserDTO;
+
 import java.util.List;
 
-import com.swapnixi.feedbackhub.payloads.UserDTO;
 
 public interface UserService {
     // DTO helps us to stop exposing the data
     UserDTO createUser(UserDTO user);
-    UserDTO updateUser(UserDTO user, Integer userId); // Fixed method name to updateUser
-    UserDTO getUserById(Integer userId); // Corrected method name to getUserById
+    UserDTO updateUser(UserDTO user, Long userId);
+    UserDTO getUserById(Long userId);
     List<UserDTO> getAllUsers();
-    void deleteUser(Integer userId);
+    void deleteUser(Long userId);
 }
